previous tests fully refactored
